+++ fabric/src/main/java/dev/nyon/magnetic/mixins/RegistryDataLoaderMixin.java
@@ -2,8 +2,8 @@ package dev.nyon.magnetic.mixins;
 
 import com.google.gson.JsonElement;
 import com.mojang.serialization.Decoder;
-import dev.nyon.magnetic.MagneticEnchantmentGeneratorKt;
 import dev.nyon.magnetic.config.ConfigKt;
+import dev.nyon.magnetic.datagen.MagneticEnchantmentGeneratorKt;
 import net.minecraft.core.RegistrationInfo;
 import net.minecraft.core.WritableRegistry;
 import net.minecraft.resources.RegistryDataLoader;