+++ fabric/src/main/kotlin/dev/nyon/magnetic/datagen/MagneticEnchantmentGenerator.kt
@@ -1,4 +1,4 @@
-package dev.nyon.magnetic
+package dev.nyon.magnetic.datagen
 
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator