+++ paper/src/main/kotlin/dev/nyon/magnetic/Config.kt
@@ -1,6 +1,10 @@
 package dev.nyon.magnetic
 
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.boolean
+import kotlinx.serialization.json.jsonObject
+import kotlinx.serialization.json.jsonPrimitive
 
 lateinit var config: Config
 
@@ -9,5 +13,19 @@ data class Config(
     var needEnchantment: Boolean = true,
     var needSneak: Boolean = false,
     var expAllowed: Boolean = true,
-    var itemsAllowed: Boolean = true
+    var itemsAllowed: Boolean = true,
+    var needPermission: Boolean = false
 )
+
+internal fun migrate(jsonElement: JsonElement, version: Int?): Config? {
+    val jsonObject = jsonElement.jsonObject
+    return when (version) {
+        1 -> Config(
+            needEnchantment = jsonObject["needEnchantment"]?.jsonPrimitive?.boolean ?: return null,
+            needSneak = jsonObject["needSneak"]?.jsonPrimitive?.boolean ?: return null,
+            expAllowed = jsonObject["expAllowed"]?.jsonPrimitive?.boolean ?: return null,
+            itemsAllowed = jsonObject["itemsAllowed"]?.jsonPrimitive?.boolean ?: return null
+        )
+        else -> null
+    }
+}