+++ changelog.md
@@ -1,2 +1,3 @@
-# Fabric
-- fix NullPointerException when trying to place a sponge in water
+# Both
+- add option to require a player to have a permission ('magnetic.ability.use') to use the ability
+  - on Fabric the fabric-permissions-api has to be installed for this to work